Reproduce run configuration bug
This is a simple app that displays its applicationId. I've committed a couple run configurations that allow us to build and install this app with application id "dev.jeran.electric.eel.bug.a" and "dev.jeran.electric.eel.bug.b". There's probably an easier way to do this, but the "app a" run configuration will first run the "configure a" script which writes the desired applicationId to app.properties. Our build script then reads this applicationId and uses it to build our app.

Here's the problem: Running app A will have us build, install, and run app A, but if we run app B, it will build and install app B, but it will run app A.

We see the following printout from the Run window:

```
02/16 13:38:18: Launching 'app b' on Pixel 3 API 30.
Install successfully finished in 564 ms.
$ adb shell am start -n "dev.jeran.electric.eel.bug.a/dev.jeran.electric.eel.bug.MainActivity" -a android.intent.action.MAIN -c android.intent.category.LAUNCHER
Connected to process 19314 on device 'Pixel_3_API_30 [emulator-5554]'.
```

Here we can see the start command is called with the wrong package name.